Moved my integration into new files that inherit from old ones
